Fixed two minor issues
- Flask does not propagate exceptions unless you tell it
- Import things that are used plz
- etc

+++ app/fabfile.py
@@ -1,4 +1,4 @@
-from fabric.contrib.files import upload_template
+from fabric.contrib.files import upload_template, exists
 from fabric.api import *
 
 env.user = "root"
@@ -8,7 +8,6 @@ def install():
     upload_template("configs/csgofort.nginx", "/etc/nginx/sites-enabled/csgofort", use_jinja=True)
     upload_template("configs/uwsgi.cmd", "/root/csgofort/app/run", use_jinja=True)
     run("chmod +x /root/csgofort/app/run")
-    run()
 
 def update():
     with cd("/root/csgofort/app/"):
@@ -22,9 +21,23 @@ def migrate():
 def restart():
     # Restart nginx
     run("kill -HUP $(cat /var/run/nginx.pid)")
-    run('kill -HUP $(cat /var/run/uwsgifort.pid)')
+
+    # Restart uwsgi (if its running)
+    if exists("/var/run/uwsgifort.pid"):
+        run('kill -HUP $(cat /var/run/uwsgifort.pid)')
 
     # TODO: restart scheduler
+
+def versions():
+    with cd("/root/csgofort/app"):
+        remote_hash = run("git rev-parse HEAD").strip()
+
+    local_hash = local("git rev-parse HEAD").strip()
+
+    if remote_hash == local_hash:
+        print "Remote and local are up-to-date!"
+    else:
+        print "%s vs %s" % (remote_hash, local_hash)
 
 def deploy():
     update()
